Extract subagent activity projection helper
